requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.3.6
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2024.10
Django==5.1.6
django-countries==7.6.1
django-crispy-forms==2.3
kombu==5.3.5
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
PyMuPDF==1.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
